backend/requirements.txt
# Flask and Extensions
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0

# Environment and Config
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Authentication
PyJWT>=2.8.0
werkzeug>=3.0.0

# GitHub Integration
PyGithub>=2.1.0

# Social Media
tweepy>=4.14.0

# AI Providers
huggingface_hub>=0.19.0
openai>=1.6.0

# Utilities
mermaid-py>=0.3.0
pyyaml>=6.0.0

# Note: API keys needed for AI providers:
# - GEMINI_API_KEY for Google Gemini
# - CLAUDE_API_KEY for Anthropic Claude
# - DEEPSEEK_API_KEY for DeepSeek
# - QWEN_API_KEY for Alibaba Qwen

# Production Server
gunicorn>=21.2.0

# Real-time Communication
flask-socketio>=5.3.6
eventlet>=0.33.3

# Virtual Environment Support
docker>=7.0.0
psutil>=5.9.0


# Async support
aiohttp>=3.9.0
nest_asyncio>=1.6.0

# Validation
pydantic>=2.0.0
